test(governance): 拆分线程消息命令入口测试
